Allow custom limitstart parameter in module helper
As part of our [K2 Content Source for YOOtheme Pro](https://github.com/joolanders/ytp-source-k2) development, we would like to be able to fetch custom k2 items. We use the content module helper for this, and we noticed that it's not possibile to set the `limitstart` parameter, which is always taken from the request. 

This PR allows for a new `limitstart` parameter to be set in the `$params` variable to achieve that, with the fallback to the request if none is set or if a specific one is present in the request.